services/land_use_regulation/requirements.txt
httpx>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
shapely>=2.0.0
pyproj>=3.0.0
